build: pom.xml for tools-deps git compatability
This commit reintroduces the pom.xml into the git repository so that
tools-deps users can use a git url for the repository.

This enables them to get the latest changes before they are published to
clojars.

It happens to also serve as a workaround to
https://github.com/metasoarous/oz/issues/78 since pulling directly from
the repository means that the uberjar artifacts won't be consumed.