fixed the way hanging comments start new blocks
If a block is only separated by a single newline from a code comment which follows it (assuming not a md-comment), then include it as part of the block. This means that if you want code-comments to be considered the start of the next block, you need to add a blank line, as you would in markdown.